[Android] Sync settings adjustments
Chromium change:
https://github.com/brave/chromium/commit/89afc5b2f860fbdb1f4dadf277ac9f4270494a68

[ManagedSyncSettings] Refactor PAYMENTS to unify more with other toggles
Starting with https://crrev.com/c/4660713, PAYMENTS is also listed as
syncer::UserSelectableType. To unify code, this patch adds the
corresponding entry to ManagedSyncSettings.mSyncTypePreferencesMap,
making it unnecessary to have a separate member variable (before this
patch, mSyncPaymentsIntegration).

Some special-casing is still needed because there is still coupling
between the PAYMENTS and AUTOFILL toggles, planned to be removed in
the future as tracked in https://crbug.com/1435431.

Bug: 1459963